clean up components a bit
